+++ test/lib/semver.test.js
@@ -1,11 +1,11 @@
 'use strict'
 
-const assert = require('assert')
+const { deepStrictEqual } = require('assert')
 const semver = require('../../lib/semver')
 
 module.exports = tests => {
-  tests.add('`semver` with a simple version.', async () => {
-    assert.deepStrictEqual(semver('1.2.3'), {
+  tests.add('`semver` with a simple version.', () => {
+    deepStrictEqual(semver('1.2.3'), {
       major: 1,
       minor: 2,
       patch: 3,
@@ -14,8 +14,8 @@ module.exports = tests => {
     })
   })
 
-  tests.add('`semver` with a complex version.', async () => {
-    assert.deepStrictEqual(semver('1.2.3-alpha.4+5'), {
+  tests.add('`semver` with a complex version.', () => {
+    deepStrictEqual(semver('1.2.3-alpha.4+5'), {
       major: 1,
       minor: 2,
       patch: 3,