+++ changelog.md
@@ -6,6 +6,7 @@
 
 - Updated dev dependencies.
 - Added a new [`hard-rejection`](https://npm.im/hard-rejection) dev dependency to ensure unhandled rejections in tests exit the process with an error.
+- Destructured `assert` imports.
 
 ## 2.0.2