+++ changelog.md
@@ -7,6 +7,7 @@
 - Updated dev dependencies.
 - Added a new [`hard-rejection`](https://npm.im/hard-rejection) dev dependency to ensure unhandled rejections in tests exit the process with an error.
 - Destructured `assert` imports.
+- Moved the `execFilePromise` helper from the `/lib` directory to `/test`, reducing the install size.
 
 ## 2.0.2